+++ mne/decoding/time_gen.py
@@ -66,7 +66,8 @@ class _GeneralizationAcrossTime(object):
     """  # noqa
     def __init__(self, picks=None, cv=5, clf=None, train_times=None,
                  test_times=None, predict_method='predict',
-                 predict_mode='cross-validation', scorer=None, n_jobs=1):
+                 predict_mode='cross-validation', scorer=None,
+                 score_mode='mean-fold-wise', n_jobs=1):
 
         from sklearn.preprocessing import StandardScaler
         from sklearn.linear_model import LogisticRegression
@@ -93,6 +94,7 @@ class _GeneralizationAcrossTime(object):
         self.clf = clf
         self.predict_mode = predict_mode
         self.scorer = scorer
+        self.score_mode = score_mode
         self.picks = picks
         self.predict_method = predict_method
         self.n_jobs = n_jobs
@@ -100,8 +102,9 @@ class _GeneralizationAcrossTime(object):
     def fit(self, epochs, y=None):
         """Train a classifier on each specified time slice.
 
-        Note. This function sets the ``picks_``, ``ch_names``, ``cv_``,
-        ``y_train``, ``train_times_`` and ``estimators_`` attributes.
+        .. note::
+            This function sets the ``picks_``, ``ch_names``, ``cv_``,
+            ``y_train``, ``train_times_`` and ``estimators_`` attributes.
 
         Parameters
         ----------
@@ -164,8 +167,8 @@ class _GeneralizationAcrossTime(object):
     def predict(self, epochs):
         """Classifiers' predictions on each specified testing time slice.
 
-        .. note:: This function sets the ``y_pred_`` and ``test_times_``
-                  attributes.
+        .. note::
+            This function sets the ``y_pred_`` and ``test_times_`` attributes.
 
         Parameters
         ----------
@@ -185,7 +188,7 @@ class _GeneralizationAcrossTime(object):
         # Check that classifier has predict_method (e.g. predict_proba is not
         # always available):
         if not hasattr(self.clf, self.predict_method):
-            raise NotImplementedError('%s does not have `%s`' % (
+            raise NotImplementedError('%s does not have "%s"' % (
                 self.clf, self.predict_method))
 
         # Check that at least one classifier has been trained
@@ -194,7 +197,7 @@ class _GeneralizationAcrossTime(object):
 
         # Check predict mode
         if self.predict_mode not in ['cross-validation', 'mean-prediction']:
-            raise ValueError('`predict_mode` must be a str, "mean-prediction" '
+            raise ValueError('predict_mode must be a str, "mean-prediction" '
                              'or "cross-validation"')
 
         # Check that training cv and predicting cv match
@@ -205,7 +208,7 @@ class _GeneralizationAcrossTime(object):
             mismatch_y = len(self.y_train_) != len(epochs)
             if heterogeneous_cv or mismatch_cv or mismatch_y:
                 raise ValueError(
-                    'When `predict_mode = "cross-validation"`, the training '
+                    'When predict_mode = "cross-validation", the training '
                     'and predicting cv schemes must be identical.')
 
         # Clean attributes
@@ -227,7 +230,7 @@ class _GeneralizationAcrossTime(object):
         elif isinstance(self.test_times, dict):
             test_times = copy.deepcopy(self.test_times)
         else:
-            raise ValueError('`test_times` must be a dict or "diagonal"')
+            raise ValueError('test_times must be a dict or "diagonal"')
 
         if 'slices' not in test_times:
             if 'length' not in self.train_times_.keys():
@@ -253,7 +256,7 @@ class _GeneralizationAcrossTime(object):
             # that the dimensionality of each estimator (i.e. training
             # time) corresponds to the dimensionality of each testing time)
             if not np.all([len(test) == len(train) for test in tests]):
-                raise ValueError('`train_times` and `test_times` must '
+                raise ValueError('train_times and test_times must '
                                  'have identical lengths')
 
         # Store all testing times parameters
@@ -312,8 +315,13 @@ class _GeneralizationAcrossTime(object):
 
         Calls ``predict()`` if it has not been already.
 
-        Note. The function updates the ``scorer_``, ``scores_``, and
-        ``y_true_`` attributes.
+        .. note::
+            The function updates the ``scorer_``, ``scores_``, and
+            ``y_true_`` attributes.
+
+        .. note::
+            If ``predict_mode`` is 'mean-prediction', ``score_mode`` is
+            automatically set to 'mean-sample-wise'.
 
         Parameters
         ----------
@@ -332,7 +340,9 @@ class _GeneralizationAcrossTime(object):
             The scores estimated by ``scorer_`` at each training time and each
             testing time (e.g. mean accuracy of ``predict(X)``). Note that the
             number of testing times per training time need not be regular;
-            else, np.shape(scores) = (n_train_time, n_test_time).
+            else, np.shape(scores) = (n_train_time, n_test_time). If
+            ``score_mode`` is 'fold-wise', np.shape(scores) = (n_train_time,
+            n_test_time, n_folds).
         """
         import sklearn.metrics
         from sklearn.base import is_classifier
@@ -352,6 +362,17 @@ class _GeneralizationAcrossTime(object):
                                    'epochs to score()')
 
         # Check scorer
+        if self.score_mode not in ('fold-wise', 'mean-fold-wise',
+                                   'mean-sample-wise'):
+            raise ValueError("score_mode must be 'fold-wise', "
+                             "'mean-fold-wise' or 'mean-sample-wise'. "
+                             "Got %s instead'" % self.score_mode)
+        score_mode = self.score_mode
+        if (self.predict_mode == 'mean-prediction' and
+                self.score_mode != 'mean-sample-wise'):
+            warn("score_mode changed from %s set to 'mean-sample-wise' because"
+                 " predict_mode is 'mean-prediction'." % self.score_mode)
+            score_mode = 'mean-sample-wise'
         self.scorer_ = self.scorer
         if self.scorer_ is None:
             # Try to guess which scoring metrics should be used
@@ -366,11 +387,11 @@ class _GeneralizationAcrossTime(object):
                 self.scorer_ = getattr(sklearn.metrics, '%s_score' %
                                        self.scorer_)
             else:
-                raise KeyError("`{0} scorer` Doesn't appear to be valid a "
+                raise KeyError("{0} scorer Doesn't appear to be valid a "
                                "scikit-learn scorer.".format(self.scorer_))
         if not self.scorer_:
-            raise ValueError('Could not find a scoring metric for `clf=%s` '
-                             ' and `predict_method=%s`. Manually define scorer'
+            raise ValueError('Could not find a scoring metric for clf=%s '
+                             ' and predict_method=%s. Manually define scorer'
                              '.' % (self.clf, self.predict_method))
 
         # If no regressor is passed, use default epochs events
@@ -409,7 +430,8 @@ class _GeneralizationAcrossTime(object):
                                 n_chunks)
         scores = parallel(p_func(
             self.y_true_, [self.y_pred_[train] for train in chunk],
-            self.scorer_) for chunk in chunks)
+            self.scorer_, score_mode, self._cv_splits)
+            for chunk in chunks)
         # TODO: np.array scores from initialization JRK
         self.scores_ = np.array([score for chunk in scores for score in chunk])
         return self.scores_
@@ -571,7 +593,7 @@ def _init_ypred(n_train, n_test, n_orig_epochs, n_dim):
     return y_pred
 
 
-def _score_slices(y_true, list_y_pred, scorer):
+def _score_slices(y_true, list_y_pred, scorer, score_mode, cv):
     """Aux function of GeneralizationAcrossTime that loops across chunks of
     testing slices.
     """
@@ -579,8 +601,19 @@ def _score_slices(y_true, list_y_pred, scorer):
     for y_pred in list_y_pred:
         scores = list()
         for t, this_y_pred in enumerate(y_pred):
-            # Scores across trials
-            scores.append(scorer(y_true, np.array(this_y_pred)))
+            if score_mode in ['mean-fold-wise', 'fold-wise']:
+                # Estimate score within each fold
+                scores_ = list()
+                for train, test in cv:
+                    scores_.append(scorer(y_true[test], this_y_pred[test]))
+                scores_ = np.array(scores_)
+                # Summarize score as average across folds
+                if score_mode == 'mean-fold-wise':
+                    scores_ = np.mean(scores_, axis=0)
+            elif score_mode == 'mean-sample-wise':
+                # Estimate score across all y_pred without cross-validation.
+                scores_ = scorer(y_true, this_y_pred)
+            scores.append(scores_)
         scores_list.append(scores)
     return scores_list
 
@@ -723,18 +756,18 @@ def _sliding_window(times, window, sfreq):
 
         if not (times[0] <= window['start'] <= times[-1]):
             raise ValueError(
-                '`start` (%.2f s) outside time range [%.2f, %.2f].' % (
+                'start (%.2f s) outside time range [%.2f, %.2f].' % (
                     window['start'], times[0], times[-1]))
         if not (times[0] <= window['stop'] <= times[-1]):
             raise ValueError(
-                '`stop` (%.2f s) outside time range [%.2f, %.2f].' % (
+                'stop (%.2f s) outside time range [%.2f, %.2f].' % (
                     window['stop'], times[0], times[-1]))
         if window['step'] < 1. / sfreq:
-            raise ValueError('`step` must be >= 1 / sampling_frequency')
+            raise ValueError('step must be >= 1 / sampling_frequency')
         if window['length'] < 1. / sfreq:
-            raise ValueError('`length` must be >= 1 / sampling_frequency')
+            raise ValueError('length must be >= 1 / sampling_frequency')
         if window['length'] > np.ptp(times):
-            raise ValueError('`length` must be <= time range')
+            raise ValueError('length must be <= time range')
 
         # Convert seconds to index
 
@@ -857,19 +890,19 @@ class GeneralizationAcrossTime(_GeneralizationAcrossTime):
     train_times : dict | None
         A dictionary to configure the training times:
 
-            ``slices`` : ndarray, shape (n_clfs,)
+            * ``slices`` : ndarray, shape (n_clfs,)
                 Array of time slices (in indices) used for each classifier.
                 If not given, computed from 'start', 'stop', 'length', 'step'.
-            ``start`` : float
+            * ``start`` : float
                 Time at which to start decoding (in seconds).
                 Defaults to min(epochs.times).
-            ``stop`` : float
+            * ``stop`` : float
                 Maximal time at which to stop decoding (in seconds).
                 Defaults to max(times).
-            ``step`` : float
+            * ``step`` : float
                 Duration separating the start of subsequent classifiers (in
                 seconds). Defaults to one time sample.
-            ``length`` : float
+            * ``length`` : float
                 Duration of each classifier (in seconds).
                 Defaults to one time sample.
 
@@ -880,7 +913,7 @@ class GeneralizationAcrossTime(_GeneralizationAcrossTime):
         each classifier is trained.
         If set to None, predictions are made at all time points.
         If set to dict, the dict should contain ``slices`` or be contructed in
-        a similar way to train_times::
+        a similar way to train_times:
 
             ``slices`` : ndarray, shape (n_clfs,)
                 Array of time slices (in indices) used for each classifier.
@@ -896,19 +929,31 @@ class GeneralizationAcrossTime(_GeneralizationAcrossTime):
         Indicates how predictions are achieved with regards to the cross-
         validation procedure:
 
-            ``cross-validation`` : estimates a single prediction per sample
+            * ``cross-validation`` : estimates a single prediction per sample
                 based on the unique independent classifier fitted in the
                 cross-validation.
-            ``mean-prediction`` : estimates k predictions per sample, based on
-                each of the k-fold cross-validation classifiers, and average
+
+            * ``mean-prediction`` : estimates k predictions per sample, based
+                on each of the k-fold cross-validation classifiers, and average
                 these predictions into a single estimate per sample.
 
-        Default: 'cross-validation'
+        Defaults to 'cross-validation'.
     scorer : object | None | str
-        scikit-learn Scorer instance or str type indicating the name of
-        the scorer such as `accuracy`, `roc_auc`. If None, set to 
-        accuracy_score.
+        scikit-learn Scorer instance or str type indicating the name of the
+        scorer such as ``accuracy``, ``roc_auc``. If None, set to ``accuracy``.
+    score_mode : {'fold-wise', 'mean-fold-wise', 'mean-sample-wise'}
+        Determines how the scorer is estimated:
 
+            * ``fold-wise`` : returns the score obtained in each fold.
+
+            * ``mean-fold-wise`` : returns the average of the fold-wise scores.
+
+            * ``mean-sample-wise`` : returns score estimated across across all
+                y_pred independently of the cross-validation. This method is
+                faster than ``mean-fold-wise`` but less conventional, use at
+                your own risk.
+
+        Defaults to 'mean-fold-wise'.
     n_jobs : int
         Number of jobs to run in parallel. Defaults to 1.
 
@@ -923,10 +968,11 @@ class GeneralizationAcrossTime(_GeneralizationAcrossTime):
     train_times_ : dict
         A dictionary that configures the training times:
 
-            ``slices`` : ndarray, shape (n_clfs,)
+            * ``slices`` : ndarray, shape (n_clfs,)
                 Array of time slices (in indices) used for each classifier.
                 If not given, computed from 'start', 'stop', 'length', 'step'.
-            ``times`` : ndarray, shape (n_clfs,)
+
+            * ``times`` : ndarray, shape (n_clfs,)
                 The training times (in seconds).
 
     test_times_ : dict
@@ -973,11 +1019,13 @@ class GeneralizationAcrossTime(_GeneralizationAcrossTime):
     """  # noqa
     def __init__(self, picks=None, cv=5, clf=None, train_times=None,
                  test_times=None, predict_method='predict',
-                 predict_mode='cross-validation', scorer=None, n_jobs=1):
+                 predict_mode='cross-validation', scorer=None,
+                 score_mode='mean-fold-wise', n_jobs=1):
         super(GeneralizationAcrossTime, self).__init__(
             picks=picks, cv=cv, clf=clf, train_times=train_times,
             test_times=test_times, predict_method=predict_method,
-            predict_mode=predict_mode, scorer=scorer, n_jobs=n_jobs)
+            predict_mode=predict_mode, scorer=scorer, score_mode=score_mode,
+            n_jobs=n_jobs)
 
     def __repr__(self):
         s = ''
@@ -1174,15 +1222,19 @@ class TimeDecoding(_GeneralizationAcrossTime):
             ``slices`` : ndarray, shape (n_clfs,)
                 Array of time slices (in indices) used for each classifier.
                 If not given, computed from 'start', 'stop', 'length', 'step'.
+
             ``start`` : float
                 Time at which to start decoding (in seconds). By default,
                 min(epochs.times).
+
             ``stop`` : float
                 Maximal time at which to stop decoding (in seconds). By
                 default, max(times).
+
             ``step`` : float
                 Duration separating the start of subsequent classifiers (in
                 seconds). By default, equals one time sample.
+
             ``length`` : float
                 Duration of each classifier (in seconds). By default, equals
                 one time sample.
@@ -1197,18 +1249,31 @@ class TimeDecoding(_GeneralizationAcrossTime):
         Indicates how predictions are achieved with regards to the cross-
         validation procedure:
 
-            ``cross-validation`` : estimates a single prediction per sample
+            * ``cross-validation`` : estimates a single prediction per sample
                 based on the unique independent classifier fitted in the
                 cross-validation.
-            ``mean-prediction`` : estimates k predictions per sample, based on
-                each of the k-fold cross-validation classifiers, and average
+
+            * ``mean-prediction`` : estimates k predictions per sample, based
+                on each of the k-fold cross-validation classifiers, and average
                 these predictions into a single estimate per sample.
 
-        Default: 'cross-validation'
+        Defaults to 'cross-validation'.
     scorer : object | None | str
-        scikit-learn Scorer instance or str type indicating the
-        name of the scorer such as `accuracy`, `roc_auc`. If None,
-        set to accuracy_score.
+        scikit-learn Scorer instance or str type indicating the name of the
+        scorer such as ``accuracy``, ``roc_auc``. If None, set to ``accuracy``.
+    score_mode : {'fold-wise', 'mean-fold-wise', 'mean-sample-wise'}
+        Determines how the scorer is estimated:
+
+            * ``fold-wise`` : returns the score obtained in each fold.
+
+            * ``mean-fold-wise`` : returns the average of the fold-wise scores.
+
+            * ``mean-sample-wise`` : returns score estimated across across all
+                y_pred independently of the cross-validation. This method is
+                faster than ``mean-fold-wise`` but less conventional, use at
+                your own risk.
+
+        Defaults to 'mean-fold-wise'.
     n_jobs : int
         Number of jobs to run in parallel. Defaults to 1.
 
@@ -1223,10 +1288,11 @@ class TimeDecoding(_GeneralizationAcrossTime):
     times_ : dict
         A dictionary that configures the training times:
 
-            ``slices`` : ndarray, shape (n_clfs,)
+            * ``slices`` : ndarray, shape (n_clfs,)
                 Array of time slices (in indices) used for each classifier.
                 If not given, computed from 'start', 'stop', 'length', 'step'.
-            ``times`` : ndarray, shape (n_clfs,)
+
+            * ``times`` : ndarray, shape (n_clfs,)
                 The training times (in seconds).
 
     cv_ : CrossValidation object
@@ -1255,13 +1321,15 @@ class TimeDecoding(_GeneralizationAcrossTime):
 
     def __init__(self, picks=None, cv=5, clf=None, times=None,
                  predict_method='predict', predict_mode='cross-validation',
-                 scorer=None, n_jobs=1):
+                 scorer=None, score_mode='mean-fold-wise', n_jobs=1):
         super(TimeDecoding, self).__init__(picks=picks, cv=cv, clf=clf,
                                            train_times=times,
                                            test_times='diagonal',
                                            predict_method=predict_method,
                                            predict_mode=predict_mode,
-                                           scorer=scorer, n_jobs=n_jobs)
+                                           scorer=scorer,
+                                           score_mode=score_mode,
+                                           n_jobs=n_jobs)
         self._clean_times()
 
     def __repr__(self):
@@ -1292,8 +1360,9 @@ class TimeDecoding(_GeneralizationAcrossTime):
     def fit(self, epochs, y=None):
         """Train a classifier on each specified time slice.
 
-        Note. This function sets the ``picks_``, ``ch_names``, ``cv_``,
-        ``y_train``, ``train_times_`` and ``estimators_`` attributes.
+        .. note::
+            This function sets the ``picks_``, ``ch_names``, ``cv_``,
+            ``y_train``, ``train_times_`` and ``estimators_`` attributes.
 
         Parameters
         ----------
@@ -1323,8 +1392,8 @@ class TimeDecoding(_GeneralizationAcrossTime):
     def predict(self, epochs):
         """Test each classifier on each specified testing time slice.
 
-        .. note:: This function sets the ``y_pred_`` and ``test_times_``
-                  attributes.
+        .. note::
+            This function sets the ``y_pred_`` and ``test_times_`` attributes.
 
         Parameters
         ----------
@@ -1350,8 +1419,13 @@ class TimeDecoding(_GeneralizationAcrossTime):
 
         Calls ``predict()`` if it has not been already.
 
-        Note. The function updates the ``scorer_``, ``scores_``, and
-        ``y_true_`` attributes.
+        .. note::
+            The function updates the ``scorer_``, ``scores_``, and
+            ``y_true_`` attributes.
+
+        .. note::
+            If ``predict_mode`` is 'mean-prediction', ``score_mode`` is
+            automatically set to 'mean-sample-wise'.
 
         Parameters
         ----------